requirements.txt
streamlit
cloudinary
openai
python-dotenv